feat: merge pipPopUP and floating bar (TNA-87)
