+++ messages/en.json
@@ -753,7 +753,8 @@
   "MiniWarningPip": {
     "getReady": "not estimating",
     "warning": "Warning!",
-    "good": "Good Posture!"
+    "good": "Good Posture!",
+    "stop": "stop"
   },
   "HelpMessageModal": {
     "header": "Measurement tip",
